+++ arch/arm/include/kernel_arch_data.h
@@ -26,7 +26,7 @@
 
 #if defined(CONFIG_CPU_CORTEX_M)
 #include <cortex_m/stack.h>
-#include <cortex_m/exc.h>
+#include <cortex_m/exception.h>
 #elif defined(CONFIG_CPU_AARCH32_CORTEX_R) || defined(CONFIG_CPU_AARCH32_CORTEX_A)
 #include <cortex_a_r/stack.h>
 #include <cortex_a_r/exception.h>
