+++ soc/arm/nxp_lpc/lpc55xxx/soc.c
@@ -18,7 +18,7 @@
 #include <zephyr/drivers/uart.h>
 #include <zephyr/linker/sections.h>
 #include <zephyr/arch/cpu.h>
-#include <cortex_m/exc.h>
+#include <cortex_m/exception.h>
 #include <fsl_power.h>
 #include <fsl_clock.h>
 #include <fsl_common.h>
